+++ hardhat.config.js
@@ -1,23 +1,21 @@
 require("@nomiclabs/hardhat-waffle");
 require("@nomiclabs/hardhat-ethers");
-require('@openzeppelin/hardhat-upgrades');
-
-// This is a sample Hardhat task. To learn how to create your own go to
-// https://hardhat.org/guides/create-task.html
-task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
-  const accounts = await hre.ethers.getSigners();
-
-  for (const account of accounts) {
-    console.log(account.address);
-  }
-});
-
-// You need to export an object to set up your config
-// Go to https://hardhat.org/config/ to learn more
+require("@openzeppelin/hardhat-upgrades");
 
 /**
  * @type import('hardhat/config').HardhatUserConfig
  */
 module.exports = {
-  solidity: "0.7.3",
+  solidity: "0.8.3",
+  networks:{
+    hardhat:{
+      chainId: 1337
+    }
+    // rinkeby: {
+		// 	url: "https://rinkeby.infura.io/v3/8669d625ab2d446c9d2600f2af74ae81", 
+		// 	accounts: {
+		// 		mnemonic: mnemonic(),
+		// 	},
+		// },
+  }
 };
